15/09/2021 13:40 Summary: Hot fix.
Files:
I. tests/utils/ReleaseModel/CnOnPn/cnonpn_prepared_release.py:
    1. Refactor class 'CnOnPnExpectedRelease':
       1.1. coefficient as integer.

+++ tests/utils/ReleaseModel/CnOnPn/cnonpn_prepared_release.py
@@ -718,7 +718,7 @@ class CnOnPnExpectedRelease:
                                     if o == "id":
                                         list_of_payload_conversions_coefficients_id.append(o)
                                     if o == "coefficient":
-                                        float(coefficient_object[o])
+                                        int(coefficient_object[o])
                         quantity_of_conversions_coefficients_object_into_payload = \
                             len(list_of_payload_conversions_coefficients_id)
                     except KeyError:
